remove dublicate code related to call banner from home and conversationVC.
